Split CloudTraceContext out into propagation/cloud_trace_context.h
